requirements.txt
python-telegram-bot[webhooks,job-queue]>=20.0
requests
pytz
psycopg2-binary
